refactor: create 'src/auth/ui' folder
This folder will hold UI related code. We have
different UI components related to auth.
