+++ src/auth/ui/vue/authForms/manageCredentials.vue
@@ -52,8 +52,8 @@
 import { PropType, defineComponent } from 'vue'
 import BaseAuthForm from './baseAuth.vue'
 import FormTitle from './formTitle.vue'
-import { SectionName, StaticProfile } from '../../types'
-import { WebviewClientFactory } from '../../../webviews/client'
+import { SectionName, StaticProfile } from '../../../credentials/types'
+import { WebviewClientFactory } from '../../../../webviews/client'
 import { AuthWebview } from '../show'
 import { AuthStatus } from './shared.vue'
