+++ src/auth/ui/vue/authForms/sharedAuthForms.css
