+++ src/auth/ui/vue/serviceItemContent/awsExplorerContent.vue
