+++ src/auth/ui/vue/serviceItemContent/baseServiceItemContent.css
