+++ src/auth/ui/vue/serviceItemContent/baseServiceItemContent.vue
