+++ src/auth/ui/vue/serviceItemContent/codeWhispererContent.vue
