+++ src/auth/ui/vue/serviceItemContent/shared.vue
