+++ src/extension.ts
@@ -10,7 +10,7 @@ import * as codecatalyst from './codecatalyst/activation'
 import { activate as activateAwsExplorer } from './awsexplorer/activation'
 import { activate as activateCloudWatchLogs } from './cloudWatchLogs/activation'
 import { initialize as initializeCredentials } from './auth/activation'
-import { initializeAwsCredentialsStatusBarItem } from './auth/statusBarItem'
+import { initializeAwsCredentialsStatusBarItem } from './auth/ui/statusBarItem'
 import { LoginManager } from './auth/deprecated/loginManager'
 import { CredentialsProviderManager } from './auth/providers/credentialsProviderManager'
 import { SharedCredentialsProviderFactory } from './auth/providers/sharedCredentialsProviderFactory'
